Revert "Admin refactor"
