--- src/views/AdminCreateVideoPost.vue
@@ -1,304 +1,0 @@
-<template>
-  <section class="px-5" id="createVideoPosts">
-    <p class="text-xl font-bold py-5">Create Video Headline</p>
-    <div class="flex flex-col gap-12">
-      <div class="" id="title">
-        <label for="" class="font-bold">Title*</label><br />
-        <input
-          type="text"
-          name="title"
-          v-model="title"
-          class="h-[2.5rem] bg-gray-100 p-1 focus:outline-none focus:border-4 focus:border-b-green-300 border-b w-full border-b-black dark:border-b-white"
-          id=""
-        />
-      </div>
-      <div class="" id="subtitle">
-        <label for="" class="font-bold">Sub Topic</label><br />
-        <input
-          type="text"
-          name="title"
-          v-model="subtitle"
-          class="h-[2.5rem] bg-gray-100 p-1 focus:outline-none focus:border-4 focus:border-b-green-300 border-b w-full border-b-black dark:border-b-white"
-          id=""
-        />
-      </div>
-      <div class="" id="mainImage">
-        <label for="article image" class="font-bold">Thumbnail*</label>
-        <figure class="w-full relative" id="article-image">
-          <div
-            after="Drag and drop images here or select from your device"
-            class="before:border-dashed before:border-4 before:border-gray-500 rounded-xl before:rounded-xl before:content-[url(@/assets/icons/svgs/admin/remix-icons/Vector.png)] after:content-[attr(after)] before:flex before:justify-center before:items-center before:w-full before:h-[12rem] before:backdrop-brightness-[.6] before:absolute z-50"
-          >
-            <img
-              v-if="file"
-              :src="image"
-              alt=""
-              class="object-cover h-[12rem] w-full rounded-xl"
-            />
-            <div v-else class="w-full h-[12rem] orunded-xl"></div>
-          </div>
-          <input
-            type="file"
-            @change="Changeimage"
-            required
-            :style="`background:url(${file}) no-repeat`"
-            class="w-full absolute opacity-0 top-0 h-full bg-center bg-contain border-black bg-gray-100 item-center justify-center"
-            name="article image"
-            accept="image/*"
-          />
-        </figure>
-        <i class="font-[500] text-blue-400 text-xs"
-          >recommended: images should be 1280px x 860px for best resolution</i
-        >
-      </div>
-
-      <div class="" id="videoContent">
-        <label for="article image" class="font-bold">Upload Video</label>
-        <figure class="w-full relative" id="article-image">
-          <!-- <div
-                after="Drag and drop images here or select from your device"
-                class="before:border-dashed before:border-4 before:border-gray-500 rounded-xl before:rounded-xl before:content-[url(@/assets/icons/svgs/admin/remix-icons/Vector.png)] after:content-[attr(after)] before:flex before:justify-center before:items-center before:w-full before:h-[12rem] before:backdrop-brightness-[.6] before:absolute z-50"
-              >
-                <img
-                  v-if="file"
-                  :src="image"
-                  alt=""
-                  class="object-cover h-[12rem] w-full rounded-xl"
-                />
-                <div v-else class="w-full h-[12rem] orunded-xl"></div>
-              </div> -->
-          <input
-            type="file"
-            @change="loadVideo"
-            required
-            :style="`background:url(${video}) no-repeat`"
-            class="w-full h-max bg-center bg-contain border-black item-center justify-center"
-            name="article image"
-            accept="video/*"
-          />
-          <video :src="thumbnail" controls></video>
-        </figure>
-      </div>
-      <div class="" id="body">
-        <label for="article image" class="font-bold">Body*</label>
-        <div class="-z-50">
-          <ckeditor
-            required
-            id="editor"
-            :editor="editor"
-            v-model="editorData"
-            :config="editorConfig"
-          />
-        </div>
-      </div>
-      <div class="" id="keyNotes">
-        <label for="summary" class="font-bold">Key Notes*</label>
-        <textarea
-          name="summary"
-          id=""
-          cols="30"
-          rows="4"
-          class="bg-gray-100 dark:bg-transparent p-1 focus:outline-none focus:border-4 focus:border-b-green-300 border-b w-full border-b-black dark:border-b-white"
-          v-model="summary"
-        ></textarea>
-      </div>
-      <div class="" id="metadata">
-        <div class="border border-gray-400 rounded-lg capitalize">
-          <p class="p-3 font-bold">Add to a Category*</p>
-          <hr class="w-full border-gray-400" />
-          <div class="lg:p-8 p-2 lg:text-md text-sm">
-            <div
-              class="my-4"
-              v-for="choice in categorySel.autocompleteItems"
-              :key="choice.slug"
-            >
-              <input
-                type="radio"
-                name="category"
-                :value="choice.slug"
-                id="category"
-                v-model="categorySel.selected"
-              />
-              <label for="category" class="lg:px-4 pl-2">{{
-                choice.cat_title
-              }}</label>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div
-        class="w-full md:flex-row items-center flex flex-col gap-4 mb-6"
-        id="buttons"
-      >
-        <div class="w-full md:w-auto flex justify-center">
-          <button
-            class="bg-white text-black border border-black rounded-lg md:rounded-sm p-3 w-full mx-auto"
-            @click="savePost()"
-          >
-            Add To Draft
-          </button>
-        </div>
-        <div class="w-full md:w-auto flex justify-center">
-          <button
-            class="bg-black dark:bg-white/60 text-white dark:text-black p-3 w-full rounded-lg md:rounded-sm mx-auto"
-            @click="publishVideo()"
-          >
-            Upload
-          </button>
-        </div>
-      </div>
-    </div>
-  </section>
-</template>
-
-<script>
-import { mapGetters } from "vuex";
-import { useToast } from "vue-toastification";
-import CKEditor from "@ckeditor/ckeditor5-vue";
-import { getAPI } from "../axios";
-import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
-
-ClassicEditor.create(document.querySelector("#snippet-classic-editor"), {
-  plugins: [
-    // SimpleUploadAdapter
-  ],
-  toolbar: {},
-  simpleUpload: {
-    uploadUrl: `${import.meta.env.VITE_API_URL}/ckeditor/upload/`,
-    withCredentials: false,
-    error: {
-      message:
-        "The image upload failed because the image was too big (max 1.5MB).",
-    },
-  },
-  // cloudServices: {
-  //   // All predefined builds include the Easy Image feature.
-  //   // Provide correct configuration values to use it.
-  //   // Read more about Easy Image - https://ckeditor.com/docs/ckeditor5/latest/features/images/image-upload/easy-image.html.
-  //   // For other image upload methods see the guide - https://ckeditor.com/docs/ckeditor5/latest/features/images/image-upload/image-upload.html.
-  // },
-})
-  .then((editor) => {
-    window.editor = editor;
-  })
-  .catch((err) => {});
-
-export default {
-  setup(props) {
-    const toast = useToast();
-    return { toast };
-  },
-  components: {
-    ckeditor: CKEditor.component,
-  },
-  data() {
-    return {
-      title: "",
-      subtitle: "",
-      editorpost: false,
-      editorData: "",
-      editor: ClassicEditor,
-      video: "",
-      editorConfig: {
-        // toolbar: ["bold", "italic", "|", "link"],
-        ckfinder: {
-          uploadUrl: `${import.meta.env.VITE_API_URL}/ckeditor/upload/`,
-          withCredentials: false,
-        },
-        // image: {
-        //   toolbar: ["toggleImageCaption", "imageTextAlternative"],
-        // },
-      },
-      image: "",
-      summary: "",
-      file: "",
-      categorySel: {
-        selected: [],
-        autocompleteItems: this.categoriesItem,
-      },
-      thumbnail: "",
-    };
-  },
-  computed: mapGetters({
-    user: "getUserState",
-  }),
-  methods: {
-    Changeimage(e) {
-      let file = e.target.files[0];
-      let imageSRC = URL.createObjectURL(file);
-      this.file = file;
-      this.image = imageSRC;
-      console.log(this.file, this.image);
-    },
-    loadVideo(e) {
-      let file = e.target.files[0];
-      let imageSRC = URL.createObjectURL(file);
-      this.video = file;
-      this.thumbnail = imageSRC;
-      console.log(this.video, this.thumbnail);
-    },
-    async publishVideo() {
-      let data = new FormData();
-      let postSlug = this.title.split(" ").join("").toLowerCase();
-      data.append("title", this.title);
-      data.append("author", this.user.slug);
-      data.append("categories", this.categorySel.selected);
-      data.append("picked", this.editorpost);
-      data.append("summary", this.summary);
-      data.append("body", this.editorData);
-      data.append("mainImage", this.file);
-      data.append("slug", postSlug);
-      data.append("sub_topic", this.subtitle);
-      data.append("video", this.video);
-      console.log(
-        postSlug,
-        data,
-        this.title.replaceAll(" ", "-").toLowerCase()
-      );
-
-      try {
-        this.toast.info("Creating Post...", {
-          timeout: false,
-          id: "login",
-        });
-        await this.$store
-          .dispatch("userModule/createPost", {
-            formData: data,
-            slug: this.user.slug,
-            postSlug: this.title.replaceAll(" ", "-").toLowerCase(),
-          })
-          .then((e) => {
-            console.log(e);
-            this.toast.dismiss("login");
-            this.toast.success("Post Successful!", {
-              timeout: 2000,
-              id: "success",
-            });
-            this.$router.go();
-          })
-          .catch((e) => {
-            console.log(e);
-          });
-      } catch (error) {
-        this.$emit("hasError", true, error.response.status);
-        setTimeout(() => {
-          this.$emit("hasError", false);
-        }, 4000);
-      }
-    },
-  },
-  created() {
-    getAPI
-      .get("/categories")
-      .then((response) => {
-        // console.log(response);
-        this.categorySel.autocompleteItems = response.data;
-      })
-      .catch((err) => {
-        console.log("Error fetching categories", err);
-      });
-  },
-};
-</script>
-
-<style lang="scss" scoped></style>
